+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Add Lambda runtime constants to the Go SDK.
+* Add module containing constants for RDS DB Instance StorageType.
 
 ## 0.18.20 (2019-07-09)
 * Fix detailed diffs with nested computed values.